requirements.txt
Flask==3.0.2
requests==2.31.0
pycryptodome==3.20.0
urllib3==2.2.1
Werkzeug==3.0.1
